Fix large GUI elements on some devices (#1848)
With high DPI scaling enabled, on some devices, the devicePixelRatio() is a bit off
from the real ratio that we used to calculate ourselves (= DPI / 160) where 160 dpi
is considered the "standard" dpi on Android. While on many phones/tablets this
difference would be small (e.g. 0-3%), on some devices this discrepancy was larger
according to our previous calculations (e.g. Huawei P30 has devicePixelRatio() = 3
while with 422 DPI it is ~2.6375 which means ~14% difference). The discrepancy
means that the whole GUI was suddenly larger than it used to be, wasting space.

So we keep using __dp to adjust scaling for these cases - under normal circumstances
__dp should be 1 or somewhere near one - in the example of Huawei P30 __dp is
~0.8774 ( = 2.6375 / 3)

In addition, record button has been replaced by SVG and few other minor GUI updates.

Fixes #1833