Add rebase files to sync (#1878)
fixes #1738

save EDIT conflict files to a project directory so that they are synced and available for users in Mergin.
Edit conflict files are generated when geodiff could not rebase local changes upon server changes.
There is also a new naming convention for conflict files. We differentiate between edit conflict (conflict in diffable file, output is json file from geodiff) and copy conflict (conflict in non diffable file or gpkg that is not possible to rebase).